Small fixes in model typing

+++ MarkUS-Backend/src/database/migrations/20210629195916-create-subject.js
@@ -46,10 +46,10 @@ module.exports = {
         onDelete: 'cascade'
       },
       avgMark: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DOUBLE
       },
       officialMark: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DOUBLE
       },
       createdAt: {
         allowNull: false,
